+++ src/main/java/com/simsilica/ethereal/NetworkStateListener.java
@@ -115,7 +115,7 @@ public class NetworkStateListener implements StateListener {
         this.zoneIndex = zoneIndex;
         this.idIndex = idIndex;
         this.space = new SharedObjectSpace(host.getObjectProtocol());
-        this.stateWriter = new StateWriter(conn, host.getObjectProtocol());        
+        this.stateWriter = new StateWriter(conn, host.getObjectProtocol(), stats);        
     }
 
     public void setSelf( Long self, Vec3d startingPosition ) {